refactoring Components
